requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
crewai>=0.22.0
langgraph>=0.0.26
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.19
playwright>=1.41.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
chromadb>=0.4.22
networkx>=3.2.1
neo4j>=5.17.0
tavily-python>=0.3.0
sse-starlette>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tiktoken>=0.6.0
nest-asyncio>=1.6.0
